Remove the received amount in favour of manually checking.
When chaining calls it can be helpful to use the amountOut received from an aggregator or swap to then use it in another function. The problem is that some aggregators do not always return the proper amountOut and it can break the chain. Instead we do all the checks and manually call `_checkBalance` when deleveraging.